+++ Views/Space/SpaceShowAllView.php
@@ -1,5 +1,8 @@
 <?php
-class SpaceShowAllView {
+include_once '../Functions/HavePermission.php';
+
+class SpaceShowAllView
+{
     private $spaces;
     private $itemsPerPage;
     private $currentPage;
@@ -7,20 +10,23 @@ class SpaceShowAllView {
     private $totalPages;
     private $stringToSearch;
 
-    function __construct($spacesData, $itemsPerPage=NULL, $currentPage=NULL, $totalSpaces=NULL, $toSearch=NULL){
+    function __construct($spacesData, $itemsPerPage = NULL, $currentPage = NULL, $totalSpaces = NULL, $toSearch = NULL)
+    {
         $this->spaces = $spacesData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalSpaces = $totalSpaces;
-        $this->totalPages = ceil($totalSpaces/$itemsPerPage);
+        $this->totalPages = ceil($totalSpaces / $itemsPerPage);
         $this->stringToSearch = $toSearch;
         $this->render();
     }
-    function render(){
+
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -30,11 +36,13 @@ class SpaceShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/SpaceController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
+                <?php else:
+                if (HavePermission("Space", "ADD")): ?>
                     <a class="btn btn-success" role="button" href="../Controllers/SpaceController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Añadir espacio"></p>
+                        <span data-feather="plus"></span>
+                        <p data-translate="Añadir espacio"></p>
                     </a>
-                <?php endif;?>
+                <?php endif; endif; ?>
 
             </div>
             <div class="table-responsive">
@@ -47,20 +55,26 @@ class SpaceShowAllView {
                         <th class="actions-row"><label data-translate="Acciones"></label></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->spaces)):?>
+                    <?php if (!empty($this->spaces)): ?>
                     <tbody>
                     <?php foreach ($this->spaces as $space): ?>
                         <tr>
-                            <td><?php echo $space->getName() ;?></td>
-                            <td><?php echo $space->getBuilding()->getName() ;?></td>
-                            <td><?php echo $space->getCapacity() ;?></td>
+                            <td><?php echo $space->getName(); ?></td>
+                            <td><?php echo $space->getBuilding()->getName(); ?></td>
+                            <td><?php echo $space->getCapacity(); ?></td>
                             <td class="row">
-                                <a href="../Controllers/SpaceController.php?action=show&id=<?php echo $space->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/SpaceController.php?action=edit&id=<?php echo $space->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/SpaceController.php?action=delete&id=<?php echo $space->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <? if (HavePermission("Space", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/SpaceController.php?action=show&id=<?php echo $space->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <? }
+                                if (HavePermission("Space", "EDIT")) { ?>
+                                    <a href="../Controllers/SpaceController.php?action=edit&id=<?php echo $space->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <? }
+                                if (HavePermission("Space", "DELETE")) { ?>
+                                    <a href="../Controllers/SpaceController.php?action=delete&id=<?php echo $space->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <? } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -72,7 +86,7 @@ class SpaceShowAllView {
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalSpaces,
-                    "Space")?>
+                    "Space") ?>
 
             </div>
         </main>
@@ -85,4 +99,5 @@ class SpaceShowAllView {
         <?php
     }
 }
+
 ?>
