+++ Views/University/UniversityShowAllView.php
@@ -1,5 +1,8 @@
 <?php
-class UniversityShowAllView {
+include_once '../Functions/HavePermission.php';
+
+class UniversityShowAllView
+{
     private $universities;
     private $itemsPerPage;
     private $currentPage;
@@ -7,20 +10,23 @@ class UniversityShowAllView {
     private $totalPages;
     private $stringToSearch;
 
-    function __construct($universitiesData, $itemsPerPage=NULL, $currentPage=NULL, $totalUniversities=NULL, $toSearch=NULL){
+    function __construct($universitiesData, $itemsPerPage = NULL, $currentPage = NULL, $totalUniversities = NULL, $toSearch = NULL)
+    {
         $this->universities = $universitiesData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalUniversities = $totalUniversities;
-        $this->totalPages = ceil($totalUniversities/$itemsPerPage);
+        $this->totalPages = ceil($totalUniversities / $itemsPerPage);
         $this->stringToSearch = $toSearch;
         $this->render();
     }
-    function render(){
+
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -30,11 +36,14 @@ class UniversityShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/UniversityController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
-                    <a class="btn btn-success" role="button" href="../Controllers/UniversityController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Añadir universidad"></p>
-                    </a>
-                <?php endif;?>
+                <?php else:
+                    if (HavePermission("University", "ADD")): ?>
+                        <a class="btn btn-success" role="button"
+                           href="../Controllers/UniversityController.php?action=add">
+                            <span data-feather="plus"></span>
+                            <p data-translate="Añadir universidad"></p>
+                        </a>
+                    <?php endif; endif; ?>
 
             </div>
             <div class="table-responsive">
@@ -46,19 +55,25 @@ class UniversityShowAllView {
                         <th class="actions-row"><label data-translate="Acciones"></label></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->universities)):?>
+                    <?php if (!empty($this->universities)): ?>
                     <tbody>
                     <?php foreach ($this->universities as $university): ?>
                         <tr>
-                            <td><?php echo $university->getName() ;?></td>
-                            <td><?php echo $university->getAcademicCourse()->getAcademicCourseAbbr() ;?></td>
+                            <td><?php echo $university->getName(); ?></td>
+                            <td><?php echo $university->getAcademicCourse()->getAcademicCourseAbbr(); ?></td>
                             <td class="row">
-                                <a href="../Controllers/UniversityController.php?action=show&id=<?php echo $university->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/UniversityController.php?action=edit&id=<?php echo $university->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/UniversityController.php?action=delete&id=<?php echo $university->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <? if (HavePermission("University", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/UniversityController.php?action=show&id=<?php echo $university->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <? }
+                                if (HavePermission("University", "EDIT")) { ?>
+                                    <a href="../Controllers/UniversityController.php?action=edit&id=<?php echo $university->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <? }
+                                if (HavePermission("University", "DELETE")) { ?>
+                                    <a href="../Controllers/UniversityController.php?action=delete&id=<?php echo $university->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <? } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -70,7 +85,7 @@ class UniversityShowAllView {
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalUniversities,
-                    "University")?>
+                    "University") ?>
 
             </div>
         </main>
@@ -83,4 +98,5 @@ class UniversityShowAllView {
         <?php
     }
 }
+
 ?>
